requirements.txt
pygame>=2.0.0
PyOpenGL>=3.1.0
PyOpenGL_accelerate>=3.1.0
numpy>=1.20.0 